refactor(llm): split entry point discovery
Move repo config and entry-point discovery helpers into skylos.llm.entry_points while keeping verify_orchestrator compatibility wrappers.

Verified with py_compile, focused orchestrator/feedback tests, full pytest, Skylos quality scan, and Skylos debt scan.
